service-ai/pyproject.toml
[project]
name = "service-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "httpx==0.27.0",
    "numpy==1.26.0",
    "openai==1.35.0",
    "pgvector==0.3.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.0",
    "pydantic-settings==2.3.0",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.0",
    "python-dotenv==1.0.0",
    "scipy==1.13.0",
    "sentence-transformers==3.0.0",
    "torch==2.3.0",
    "uvicorn[standard]==0.30.0",
]
